SME-35 remove some auth abstractions (still working only with keycloak auth)
